+++ tests/e2e/features/disable_remote_media.feature
@@ -1,21 +1,21 @@
 Feature: Remote media disabling
 
-  Scenario: Remote video Track stops on disable
+  Scenario: Remote video Track stops when disabled
     Given room with joined member Alice and Bob
     When  Alice disables remote video
     Then Alice remote device video track from Bob disables
 
-  Scenario: Remote audio Track stops on disable
+  Scenario: Remote audio Track stops when disabled
     Given room with joined member Alice and Bob
     When  Alice disables remote audio
     Then Alice remote audio track from Bob disables
 
-  Scenario: Remote Track.on_disabled fires on disable audio
+  Scenario: RemoteTrack.on_disabled fires on disable audio
     Given room with joined member Alice and Bob
     When Alice disables remote audio
     Then on_disabled callback fires 1 time on Alice's remote audio track from Bob
 
-  Scenario: Remote Track.on_disabled fires on disable video
+  Scenario: RemoteTrack.on_disabled fires on disable video
     Given room with joined member Alice and Bob
     When Alice disables remote video
     Then on_disabled callback fires 1 time on Alice's remote device video track from Bob
