+++ tests/e2e/features/local_tracks_create.feature
@@ -1,6 +1,6 @@
 Feature: Local Track are created
 
-  Scenario: Local Track are created on connect
+  Scenario: Local Tracks are created on connect
     Given room with member Alice
     And joined member Bob
     When Alice joins room
@@ -8,20 +8,20 @@ Feature: Local Track are created
     And Alice has local device video
     And Alice has local audio
 
-  Scenario: Local Tracks are not created when all media disabled
+  Scenario: Local Tracks are not created if all media is disabled
     Given room with member Alice with disabled media publishing
     And joined member Bob
     When Alice joins room
     Then Member Alice has 0 local tracks
 
-  Scenario: Local Track creates when Member enables video
+  Scenario: Local video Track is created when Member enables video
     Given room with joined member Alice with disabled media publishing
     And joined member Bob
     When Alice enables video
     Then Member Alice has 1 local tracks
     And Alice has local device video
 
-  Scenario: Local Track creates when Member enables audio
+  Scenario: Local audio Track is created when Member enables audio
     Given joined member Alice with disabled media publishing
     And joined member Bob
     When Alice enables audio
