+++ tests/e2e/world/member.rs
@@ -12,29 +12,32 @@ use crate::{
     },
 };
 
-/// All errors which can happen while working with [`Member`].
+/// All errors which can happen while working with a [`Member`].
 #[derive(Debug, Display, Error, From)]
 pub enum Error {
-    /// [`Room`] or [`ConnectionStore`] object errored.
+    /// [`Room`] or a [`ConnectionStore`] object errored.
     Object(object::Error),
 }
 
+/// Shortcut for a [`Result`] containing an [`Error`](enum@Error).
+///
+/// [`Result`]: std::result::Result
 type Result<T> = std::result::Result<T, Error>;
 
-/// Builder for the [`Member`].
-pub struct MemberBuilder {
-    /// ID with which [`Member`] will be created.
+/// Builder of a [`Member`].
+pub struct Builder {
+    /// ID with which a [`Member`] will be created.
     pub id: String,
 
-    /// Flag which indicates that [`Member`] will publish media.
+    /// Indicator whether a [`Member`] will publish media.
     pub is_send: bool,
 
-    /// Flag which indicates that [`Member`] will receive media.
+    /// Indicator whether a [`Member`] will receive media.
     pub is_recv: bool,
 }
 
-impl MemberBuilder {
-    /// Creates new [`Member`] with a [`MemberBuilder`] configuration.
+impl Builder {
+    /// Creates a new [`Member`] out of this [`Builder`] configuration.
     pub async fn build(self, room: Object<Room>) -> Result<Member> {
         let connection_store = room.connections_store().await?;
         let mut media_state = HashMap::new();
@@ -54,38 +57,39 @@ impl MemberBuilder {
     }
 }
 
-/// Object which represents some connected to the Media Server `Member`.
+/// [`Object`] representing a `Member` connected to a media server.
 pub struct Member {
-    /// ID of [`Member`] on the Media Server.
+    /// ID of this [`Member`] on a media server.
     id: String,
 
-    /// Flag which indicates that [`Member`] should publish media.
+    /// Indicator whether this [`Member`] should publish media.
     is_send: bool,
 
-    /// Flag which indicates that [`Member`] should receive media.
+    /// Indicator whether this [`Member`] should receive media.
     is_recv: bool,
 
-    /// Flag which indicates that [`Member`] is joined to the `Room`.
+    /// Indicator whether this [`Member`] is joined a [`Room`] on a media
+    /// server.
     is_joined: bool,
 
-    /// Publishing media state of this [`Member`].
+    /// Media publishing state of this [`Member`].
     ///
     /// If value is `true` then this [`MediaKind`] and [`MediaSourceKind`] is
     /// enabled.
     send_state: RefCell<HashMap<(MediaKind, MediaSourceKind), bool>>,
 
-    /// Receiving media state of this [`Member`].
+    /// Media receiving state of this [`Member`].
     ///
     /// If value is `true` then this [`MediaKind`] and [`MediaSourceKind`] is
     /// enabled.
     recv_state: RefCell<HashMap<(MediaKind, MediaSourceKind), bool>>,
 
-    /// Representation of the `Room` JS object.
+    /// [`Room`]'s [`Object`] that this [`Member`] is intended to join.
     room: Object<Room>,
 
-    /// Storage for the [`Connection`]s throws by this [`Member`]'s `Room`.
+    /// Storage of [`Connection`]s thrown by this [`Member`]'s [`Room`].
     ///
-    /// [`Connection`]: crate::object::connection::Connection
+    /// [`Connection`]: object::connection::Connection
     connection_store: Object<ConnectionStore>,
 }
 
@@ -133,32 +137,6 @@ impl Member {
         Ok(())
     }
 
-    /// Returns list of [`MediaKind`]s and [`MediaSourceKind`] based on the
-    /// provided [`Option`]s.
-    fn kinds_and_source_kinds(
-        kind: Option<MediaKind>,
-        source_kind: Option<MediaSourceKind>,
-    ) -> Vec<(MediaKind, MediaSourceKind)> {
-        let mut kinds_and_source_kinds = Vec::new();
-        if let Some(kind) = kind {
-            if let Some(source_kind) = source_kind {
-                kinds_and_source_kinds.push((kind, source_kind));
-            } else {
-                kinds_and_source_kinds.push((kind, MediaSourceKind::Device));
-            }
-        } else if let Some(source_kind) = source_kind {
-            kinds_and_source_kinds.push((MediaKind::Audio, source_kind));
-            kinds_and_source_kinds.push((MediaKind::Video, source_kind));
-        } else {
-            kinds_and_source_kinds
-                .push((MediaKind::Video, MediaSourceKind::Device));
-            kinds_and_source_kinds
-                .push((MediaKind::Audio, MediaSourceKind::Device));
-        }
-
-        kinds_and_source_kinds
-    }
-
     /// Updates [`Member::send_state`].
     fn update_media_state(
         &self,
@@ -166,9 +144,7 @@ impl Member {
         source_kind: Option<MediaSourceKind>,
         enabled: bool,
     ) {
-        for (kind, source_kind) in
-            Self::kinds_and_source_kinds(kind, source_kind)
-        {
+        for (kind, source_kind) in kinds_combinations(kind, source_kind) {
             *self
                 .send_state
                 .borrow_mut()
@@ -184,9 +160,7 @@ impl Member {
         source_kind: Option<MediaSourceKind>,
         enabled: bool,
     ) {
-        for (kind, source_kind) in
-            Self::kinds_and_source_kinds(kind, source_kind)
-        {
+        for (kind, source_kind) in kinds_combinations(kind, source_kind) {
             *self
                 .recv_state
                 .borrow_mut()
@@ -333,12 +307,42 @@ impl Member {
     /// [`Member`]'s [`Room`].
     ///
     /// [`Connection`]: object::connection::Connection
+    #[inline]
+    #[must_use]
     pub fn connections(&self) -> &Object<ConnectionStore> {
         &self.connection_store
     }
 
     /// Returns reference to the [`Room`] of this [`Member`].
+    #[inline]
+    #[must_use]
     pub fn room(&self) -> &Object<Room> {
         &self.room
     }
+}
+
+/// Returns list of [`MediaKind`]s and [`MediaSourceKind`] based on the
+/// provided [`Option`]s.
+fn kinds_combinations(
+    kind: Option<MediaKind>,
+    source_kind: Option<MediaSourceKind>,
+) -> Vec<(MediaKind, MediaSourceKind)> {
+    let mut kinds_and_source_kinds = Vec::new();
+    if let Some(kind) = kind {
+        if let Some(source_kind) = source_kind {
+            kinds_and_source_kinds.push((kind, source_kind));
+        } else {
+            kinds_and_source_kinds.push((kind, MediaSourceKind::Device));
+        }
+    } else if let Some(source_kind) = source_kind {
+        kinds_and_source_kinds.push((MediaKind::Audio, source_kind));
+        kinds_and_source_kinds.push((MediaKind::Video, source_kind));
+    } else {
+        kinds_and_source_kinds
+            .push((MediaKind::Video, MediaSourceKind::Device));
+        kinds_and_source_kinds
+            .push((MediaKind::Audio, MediaSourceKind::Device));
+    }
+
+    kinds_and_source_kinds
 }
